chore: 409 - move immersion feature tests to specific folders
