+++ test/feature/immersion/immersion_details_test.dart
@@ -9,9 +9,9 @@ import 'package:pass_emploi_app/redux/states/favoris_state.dart';
 import 'package:pass_emploi_app/repositories/immersion_details_repository.dart';
 import 'package:pass_emploi_app/repositories/offre_emploi_details_repository.dart';
 
-import '../doubles/dummies.dart';
-import '../doubles/fixtures.dart';
-import '../utils/test_setup.dart';
+import '../../doubles/dummies.dart';
+import '../../doubles/fixtures.dart';
+import '../../utils/test_setup.dart';
 
 main() {
   test("immersion should be loaded and result displayed", () async {
